SHRINKRES-257 EmbeddedMaven-refactoring-created separated packages for maven.equipped and pom.equipped related classes and interfaces
