+++ maven/api-maven-embedded/src/main/java/org/jboss/shrinkwrap/resolver/api/maven/embedded/pom/equipped/ConfigurationDistributionStage.java
@@ -1,4 +1,6 @@
-package org.jboss.shrinkwrap.resolver.api.maven.embedded;
+package org.jboss.shrinkwrap.resolver.api.maven.embedded.pom.equipped;
+
+import org.jboss.shrinkwrap.resolver.api.maven.embedded.DistributionStage;
 
 /**
  * @author <a href="mailto:mjobanek@redhat.com">Matous Jobanek</a>
