+++ maven/impl-maven-embedded/src/main/java/org/jboss/shrinkwrap/resolver/impl/maven/embedded/invoker/equipped/MavenInvokerUnequippedEmbeddedMavenImpl.java
@@ -1,9 +1,9 @@
-package org.jboss.shrinkwrap.resolver.impl.maven.embedded;
+package org.jboss.shrinkwrap.resolver.impl.maven.embedded.invoker.equipped;
 
 import org.apache.maven.shared.invoker.InvocationRequest;
 import org.apache.maven.shared.invoker.Invoker;
-import org.jboss.shrinkwrap.resolver.api.maven.embedded.MavenInvokerEquippedEmbeddedMaven;
-import org.jboss.shrinkwrap.resolver.api.maven.embedded.MavenInvokerUnequippedEmbeddedMaven;
+import org.jboss.shrinkwrap.resolver.api.maven.embedded.invoker.equipped.MavenInvokerEquippedEmbeddedMaven;
+import org.jboss.shrinkwrap.resolver.api.maven.embedded.invoker.equipped.MavenInvokerUnequippedEmbeddedMaven;
 
 /**
  * @author <a href="mailto:mjobanek@redhat.com">Matous Jobanek</a>
