+++ maven/impl-maven-embedded/src/main/java/org/jboss/shrinkwrap/resolver/impl/maven/embedded/pom/equipped/ResolverErrorOutputHandler.java
@@ -1,21 +1,21 @@
-package org.jboss.shrinkwrap.resolver.impl.maven.embedded;
+package org.jboss.shrinkwrap.resolver.impl.maven.embedded.pom.equipped;
 
 import org.apache.maven.shared.invoker.InvocationOutputHandler;
 
 /**
  * @author <a href="mailto:mjobanek@redhat.com">Matous Jobanek</a>
  */
-public class ResolverOutputHandler implements InvocationOutputHandler {
+public class ResolverErrorOutputHandler implements InvocationOutputHandler {
 
     private final StringBuffer logBuffer;
 
-    public ResolverOutputHandler(StringBuffer logBuffer) {
+    public ResolverErrorOutputHandler(StringBuffer logBuffer) {
         this.logBuffer = logBuffer;
     }
 
     @Override
     public void consumeLine(String line) {
-        System.out.println("-> " + line);
+        System.err.println("-> " + line);
         logBuffer.append(line).append("\n");
     }
 }
