+++ maven/impl-maven-embedded/src/main/java/org/jboss/shrinkwrap/resolver/impl/maven/embedded/pom/equipped/ResolverOutputHandler.java
@@ -1,21 +1,21 @@
-package org.jboss.shrinkwrap.resolver.impl.maven.embedded;
+package org.jboss.shrinkwrap.resolver.impl.maven.embedded.pom.equipped;
 
 import org.apache.maven.shared.invoker.InvocationOutputHandler;
 
 /**
  * @author <a href="mailto:mjobanek@redhat.com">Matous Jobanek</a>
  */
-public class ResolverErrorOutputHandler implements InvocationOutputHandler {
+public class ResolverOutputHandler implements InvocationOutputHandler {
 
     private final StringBuffer logBuffer;
 
-    public ResolverErrorOutputHandler(StringBuffer logBuffer) {
+    public ResolverOutputHandler(StringBuffer logBuffer) {
         this.logBuffer = logBuffer;
     }
 
     @Override
     public void consumeLine(String line) {
-        System.err.println("-> " + line);
+        System.out.println("-> " + line);
         logBuffer.append(line).append("\n");
     }
 }
