--- maven/impl-maven-embedded/src/main/resources/META-INF/services/org.jboss.shrinkwrap.resolver.api.maven.embedded.MavenInvokerUnequippedEmbeddedMaven
@@ -1,1 +1,0 @@
-org.jboss.shrinkwrap.resolver.impl.maven.embedded.MavenInvokerUnequippedEmbeddedMavenImpl
