+++ maven/impl-maven-embedded/src/main/resources/META-INF/services/org.jboss.shrinkwrap.resolver.api.maven.embedded.invoker.equipped.MavenInvokerUnequippedEmbeddedMaven
@@ -1,0 +1,1 @@
+org.jboss.shrinkwrap.resolver.impl.maven.embedded.invoker.equipped.MavenInvokerUnequippedEmbeddedMavenImpl
